+++ src/components/calendar/EventCard.tsx
@@ -1,9 +1,10 @@
 import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
-import { Ionicons, FontAwesome } from '@expo/vector-icons';
+import { Ionicons } from '@expo/vector-icons';
 import { useTranslation } from 'react-i18next';
 import colors from '../../constants/colors';
 import type { CalendarEvent } from '../../api/calendar';
-import calStyles from '../../styles/calendar/calendar';
+import calStyles from './styles';
+import ChecklistSection from './ChecklistSection';
 
 interface EventCardProps {
   event: CalendarEvent;
@@ -64,34 +65,11 @@ const EventCard = ({ event, expanded, isPast, onToggleExpand, onToggleCheck }: E
       </View>
 
       {expanded && checklistItems.length > 0 && (
-        <View style={[calStyles.checklistWrap, { backgroundColor: `${event.calendarColor}26` }]}>
-          {checklistItems.map((item) => (
-            <TouchableOpacity
-              key={item.checklistId}
-              style={calStyles.checkItem}
-              onPress={() => onToggleCheck(item.checklistId)}
-              activeOpacity={0.7}
-              accessibilityRole="checkbox"
-              accessibilityLabel={item.content}
-              accessibilityState={{ checked: item.isCompleted }}
-            >
-              <View
-                style={[
-                  calStyles.checkbox,
-                  { borderColor: event.calendarColor },
-                  item.isCompleted && { backgroundColor: event.calendarColor },
-                ]}
-              >
-                {item.isCompleted && (
-                  <FontAwesome name="check" size={10} color={colors.text.white} />
-                )}
-              </View>
-              <Text style={[calStyles.checkText, item.isCompleted && calStyles.checkTextDone]}>
-                {item.content}
-              </Text>
-            </TouchableOpacity>
-          ))}
-        </View>
+        <ChecklistSection
+          items={checklistItems}
+          calendarColor={event.calendarColor}
+          onToggleCheck={onToggleCheck}
+        />
       )}
     </View>
   );
